feat(write): zip using uzip
