+++ lib/xlsx/xlsx.js
@@ -1,10 +1,8 @@
 const fs = require('fs');
 const JSZip = require('jszip');
 const {PassThrough} = require('readable-stream');
-const ZipStream = require('../utils/zip-stream');
+const UZip = require('../utils/u-zip');
 const StreamBuf = require('../utils/stream-buf');
-const ClientZip = require('../utils/client-zip');
-
 const utils = require('../utils/utils');
 const XmlStream = require('../utils/xml-stream');
 const {bufferToString} = require('../utils/browser-buffer-decode');
@@ -643,8 +641,7 @@ class XLSX {
   async write(stream, options) {
     options = options || {};
     const {model} = this.workbook;
-    // const zip = new ZipStream.ZipWriter(options.zip);
-    const zip = new ClientZip.ClientZipWriter();
+    const zip = new UZip.UZipWriter();
     zip.pipe(stream);
 
     this.prepareModel(model, options);
