+++ package.json
@@ -97,7 +97,6 @@
   ],
   "dependencies": {
     "archiver": "^5.0.0",
-    "client-zip": "^2.0.0",
     "dayjs": "^1.8.34",
     "fast-csv": "^4.3.1",
     "jszip": "^3.7.1",
@@ -105,7 +104,8 @@
     "saxes": "^5.0.1",
     "tmp": "^0.2.0",
     "unzipper": "^0.10.11",
-    "uuid": "^8.3.0"
+    "uuid": "^8.3.0",
+    "uzip": "^0.20201231.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.10.5",
